+++ R/survatr-package.R
@@ -1,6 +1,8 @@
 #' @keywords internal
 #' @importFrom data.table := .N .SD data.table is.data.table as.data.table copy setkeyv shift
 #' @importFrom rlang abort warn inform
+#' @importFrom numDeriv jacobian
+#' @importFrom sandwich bread estfun
 "_PACKAGE"
 
 ## usethis namespace: start